Adjust how workers are run.
Use prefetch-multiplier=1, -Ofair, and set tasks to acks_late and
reject_on_worker_lost.  This makes teh first available worker run a
task, and tasks aren't allocated until a worker is ready.  If a worker
fails (or shutsdown), the task will be rerun on another worker rather
than discarded.  These may be the wrong settings for workers that have
intermittent connectivity.

This also exposes a command line flag for deploy_docker of
--worker-api-url which is the url of how workers reach the Girder
instance.  This only needs to be passed to the Girder container (not the
worker containers).